design: modal 수정
