+++ src/components/auctionPostPage/Modal/index.module.scss
@@ -7,6 +7,8 @@
   background-color: rgba(0,0,0,0.5);
   top:0; 
   left: 0;
+  margin: 0;
+  z-index: 100;
 }
 
 .modal {
@@ -25,7 +27,6 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  //align-items: center;
 
   label {
     margin-left: 7rem;
@@ -40,6 +41,11 @@
     border-radius: 5px;
     font-size: 1.6rem;
     margin-left: 7rem;
+    padding-left: 10px;
+
+    &:focus {
+      outline: 0.1rem solid var(--main-color);
+    }
   }
   
   button {
@@ -53,4 +59,17 @@
     font-size: 1.6rem;
     margin-left: 7rem;
   }
+}
+
+.linkInputContainer {
+  position: relative;
+}
+
+.icon {
+  position: absolute;
+  width: 15px;
+  height: 15px;
+  top: calc(50% - 12px);
+  right: 95px;
+  color: gray;
 }
