+++ src/components/auctionPostPage/ProductInfo/index.module.scss
