+++ src/components/auctionPostPage/ProductInfo/index.tsx
