+++ src/components/auctionPostPage/ProductInfoDetail/index.module.scss
@@ -28,7 +28,6 @@
     display: flex;
     flex-direction: row;
     align-items: center;
-    margin-bottom: 1rem;
 
     span {
       font-size: 2rem;
