+++ src/components/auctionPostPage/ProductInfoDetail/index.tsx
