+++ src/components/auctionPostPage/ProductInfoImageAndName/index.module.scss
