+++ src/components/auctionPostPage/ProductInfoImageAndName/index.tsx
